pyproject.toml
[tool.poetry]
name = "lothbrok"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^1.53"
GitPython = "^3.1.11"
docker = "^4.3.1"
Flask = "^1.1.2"
semantic-version = "^2.8.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
pylint = "^2.6.0"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "processor",
    "search"
]
